+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.18
 
 - Add Update API command
+- Add support for Tools navigation
 
 ## 0.17